web: Common: Structured files of GroupSelector
