+++ web/ASC.Web.Common/src/components/group-selector/GroupSelector.stories.js
